# openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-Spend API
  description: |
    Stellar-Spend converts Stellar USDC to fiat currency (off-ramp) using the
    Allbridge cross-chain bridge and Paycrest fiat settlement layer.

    ## Authentication

    | Route group | Auth method |
    |---|---|
    | `/api/offramp/*` (legacy) | None — server-side secrets only |
    | `/api/v1/*` | API key via `X-API-Key` or `Authorization: Bearer` |
    | `/api/api-keys/*` | Admin bearer token via `Authorization: Bearer <API_KEY_ADMIN_TOKEN>` |
    | `/api/auth/2fa/*` | Public — user-supplied credentials in body |
    | `/api/webhooks/paycrest` | HMAC-SHA256 via `X-Paycrest-Signature` |

    ## Versioning

    All routes are available under `/api/v1/`. Legacy `/api/` routes are deprecated
    (sunset `2026-01-01`) and return `Deprecation`, `Sunset`, and `Link` headers.

    Version can also be negotiated with:
    - `X-API-Version: 1`
    - `Accept: application/vnd.stellarspend.v1+json`

    ## Idempotency

    Mutating endpoints support the `Idempotency-Key` header. Replay responses
    include `Idempotency-Status: replayed`; conflicts return `409`.

  version: "1.0.0"
  contact:
    name: Stellar-Spend
    url: https://github.com/Lex-Studios/Stellar-Spend
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://your-domain.com
    description: Production
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Programmatic API key for `/api/v1/*` endpoints
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        For `/api/v1/*` endpoints use a programmatic API key.
        For `/api/api-keys/*` admin endpoints use `API_KEY_ADMIN_TOKEN`.
    PaycrestSignature:
      type: apiKey
      in: header
      name: X-Paycrest-Signature
      description: HMAC-SHA256 hex digest of the raw request body (webhook only)

  # ───────── Reusable schemas ─────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Human-readable error message
...
    Currency:
      type: object
      properties:
        code:
          type: string
          example: NGN
        name:
          type: string
          example: Nigerian Naira
        symbol:
          type: string
          example: "₦"
...
    SubmitSorobanRequest:
      type: object
      required: [signedXdr]
      properties:
        signedXdr:
          type: string
          example: "AAAAAgAAAA..."
    TxStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, SUCCESS, FAILED, NOT_FOUND]
          example: PENDING
        hash:
          type: string
          example: abc123def456
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          example: uuid
        name:
          type: string
          example: Production Partner
        keyPrefix:
          type: string
          example: abc123def456
        status:
          type: string
          enum: [active, revoked, rotated]
          example: active
        rateLimitMaxRequests:
          type: integer
          example: 120
        rateLimitWindowMs:
          type: integer
          example: 60000
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Production Partner
        rateLimitMaxRequests:
          type: integer
          example: 120
        rateLimitWindowMs:
          type: integer
          example: 60000
    CreateApiKeyResponse:
      type: object
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/ApiKey"
            - type: object
              properties:
                plaintextKey:
                  type: string
                  example: ssp_live_abc123def456.very-secret-value
...
paths:
...
  /api/offramp/currencies:
    get:
      tags: [offramp]
      summary: List supported fiat currencies
      operationId: getCurrencies
      description: Returns supported currencies. Response is cached for 5 minutes.
      responses:
        "200":
          description: List of currencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Currency"
              example:
                data:
                  - code: NGN
                    name: Nigerian Naira
                    symbol: "₦"
                  - code: KES
                    name: Kenyan Shilling
                    symbol: KSh
        "500":
          description: Paycrest API unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/offramp/currencies:
    get:
      tags: [offramp]
      summary: List supported fiat currencies (v1)
      operationId: getCurrenciesV1
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      responses:
        "200":
          description: List of currencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Currency"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Paycrest API unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Institutions ─────────────────────────────────────────────────
...
  /api/offramp/bridge/submit-soroban:
    post:
      tags: [bridge]
      summary: Submit a signed Soroban XDR transaction
      operationId: submitSoroban
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitSorobanRequest"
      responses:
        "200":
          description: Transaction accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxStatusResponse"
        "400":
          description: Missing XDR, RPC error, or transaction rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Soroban RPC not configured or unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/offramp/bridge/submit-soroban:
    post:
      tags: [bridge]
      summary: Submit a signed Soroban XDR transaction (v1)
      operationId: submitSorobanV1
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitSorobanRequest"
      responses:
        "200":
          description: Transaction accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxStatusResponse"
        "400":
          description: RPC error or transaction rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Bridge: TX Status (Soroban) ───────────────────────────────────
...
  /api/api-keys:
    get:
      tags: [api-keys]
      summary: List all API keys
      operationId: listApiKeys
      security:
        - BearerAuth: []
      description: Requires `API_KEY_ADMIN_TOKEN` as bearer token.
      responses:
        "200":
          description: List of API keys (no plaintext secrets)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
        "401":
          description: Invalid or missing admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [api-keys]
      summary: Create a new API key
      operationId: createApiKey
      security:
        - BearerAuth: []
      description: |
        Requires `API_KEY_ADMIN_TOKEN` as bearer token.
        The plaintext key is returned **once** and cannot be retrieved again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...